+++ .reaxpress/helpers/react/templates/component.jsx
@@ -1,5 +1,6 @@
 module.exports = component =>
 `import React from 'react';
+import PropTypes from 'prop-types';
 import Reaxpress from '../../Reaxpress';
 import Header from '../_global/Header';
 import Footer from '../_global/Footer';
@@ -25,7 +26,7 @@ ${component}.defaultProps = {
 };
 
 ${component}.propTypes = {
-  reaxpressData: React.PropTypes.object,
+  reaxpressData: PropTypes.object,
 };
 
 export default ${component};