+++ src/react/App/Admin/Pages/Update/index.jsx
@@ -1,5 +1,6 @@
 import validator from 'validator';
 import React from 'react';
+import PropTypes from 'prop-types';
 import Reaxpress from '../../../../Reaxpress';
 import Header from '../../../_global/Header';
 import Footer from '../../../_global/Footer';
@@ -63,7 +64,7 @@ AdminPagesUpdate.defaultProps = {
 };
 
 AdminPagesUpdate.propTypes = {
-  reaxpressData: React.PropTypes.object,
+  reaxpressData: PropTypes.object,
 };
 
 export default AdminPagesUpdate;